[14.0][IMP] fielservice_geoengine: proposal for refactor of fieldservice_geoengine module.
WARNING: this PR was made some time ago. I fixed some things and had to rebase on the most recent commit.

I would like to propose some improvements to the fieldservice_geoengine module that we have been using and that we think could be helpful to the OCA field-service community.

 ha been changed to be related to  with a many2one. This removes the need for a local shape and instead uses an existing location.
By doing so some methods are not needed anymore because they are in  already.

 has also been subject to changes:
 - some missing fields have been added.
 -  attribute has been altered to a computed field in order to avoid unnecessary recalculations when it's requested.
It's recomputed only when partner's latitude and/or longitude are changed.

The map view previously added by this module in the fsm_order form view has also been made available in fsm_location form view. This change was made so that a user does not have to generate an fsm_order in order to see the map with the fsm_location's position.

In addition to the previous changes the tests have been expanded for a broader control.